python/requirements.txt
PyJWT
Flask
PyMySQL
python-dotenv
bcrypt
smtplib
flask-cors